insure filename is defined and exists before calling unlink
svn path=/bioperl-live/trunk/; revision=4890

+++ Bio/Seq/LargePrimarySeq.pm
@@ -285,7 +285,7 @@ sub DESTROY {
     my $fh = $self->_fh();
     close($fh) if( defined $fh );
     # this should be handled by Tempfile removal, but we'll unlink anyways.
-    unlink $self->_filename();
+    unlink $self->_filename() if defined $self->_filename() && -e $self->_filename;
     $self->SUPER::DESTROY();
 }
